+++ src/app/app.module.ts
@@ -1,13 +1,14 @@
+import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
-import { RouterModule, Routes } from '@angular/router';
 import { SimpleNotificationsModule } from 'angular2-notifications';
 import { SelectModule } from 'ng2-select';
 
 import { AppComponent } from './app.component';
+import { AppRoutingModule } from './app-routing.module';
+
 import { NavbarComponent } from './components/navbar/navbar.component';
 import { DeedsComponent } from './components/deeds/deeds.component';
 import { AddDeedComponent } from './components/add-deed/add-deed.component';
@@ -21,17 +22,6 @@ import { DeedService } from './services/deed.service';
 import { SchemaService } from './services/schema.service';
 import { SearchService } from './services/search.service';
 import { PagerService } from './services/pager.service';
-
-
-const appRoutes: Routes = [
-  {path: '', component:DeedsComponent},
-  {path: 'deed/add', component:AddDeedComponent},
-  {path: 'deed/:id', component:DeedDetailsComponent},
-  {path: 'deed/edit/:id', component:EditDeedComponent},
-  {path: 'schema', component:SchemaComponent},
-  {path: 'search', component:SearchComponent},
-
-]
 
 
 @NgModule({
@@ -52,7 +42,7 @@ const appRoutes: Routes = [
     FormsModule,
     ReactiveFormsModule,
     HttpModule,
-    RouterModule.forRoot(appRoutes),
+    AppRoutingModule,
     SelectModule
   ],
   providers: [DeedService, SchemaService, SearchService, PagerService],
